conf-fs: skip config if rbd-usb-run-conf.flag is set
Users can now control whether or not the conf FS should be exposed prior
to RBD mapping. This makes it possible to proceed to RBD mapping without
needing a SCSI initiator capable of sending an eject request - Android
doesn't appear to be able to do this for LU's flagged as removable.

Signed-off-by: David Disseldorp <ddiss@suse.de>